moved tests into separate files
