+++ .gitignore
@@ -2,3 +2,4 @@ outputs
 node_modules
 awsConfig.json
 .DS_Store
+test/outputs
