+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "data-archiver",
   "version": "0.0.1",
-  "description": "Utility for bulk transeferral of NGDS data onto Amazon S3.",
+  "description": "Utility for bulk transferral of NGDS data onto Amazon S3.",
   "main": "index.js",
   "dependencies": {
     "yargs": "1.2.1",
@@ -25,7 +25,7 @@
     "npm": "1.3.x"
   },
   "scripts": {
-    "test": "nodeunit testing/index.js"
+    "test": "nodeunit test/index.js"
   },
   "repository": {
     "type": "git",
