+++ test/index.js
@@ -1,0 +1,8 @@
+// Run all tests (uses nodeunit)
+
+// Run all tests
+module.exports = {
+	"test_utility": require("./test_utility"),
+	"test_parse": require("./test_parse"),
+	"test_other": require("./test_other")
+}
